Combined the scenes, problems
The UI integration is completed for the combined scenes, but placing measurement points and measuring in general does not work yet.
Also, the measurement controller (Button State) will need to be updated to allow the measurement points to attach to the boiler.
Also, ideally, the measurement points should probably STAY attached to the boiler. (Update with Boiler transform)